# swagger.json
{"openapi":"3.0.0","paths":{"/api/user/update-user":{"post":{"operationId":"updateUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserResponse"}}}}},"tags":["user"]}},"/api/auth/signup":{"post":{"operationId":"signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupResponse"}}}}},"tags":["auth"]}},"/api/auth/login":{"post":{"operationId":"login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}},"tags":["auth"]}},"/api/auth/change-password":{"patch":{"operationId":"changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["auth"]}},"/api/list":{"post":{"operationId":"addList","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddListDTO"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/List"}}}}}},"tags":["list"]}},"/api/list/{id}/{owner}":{"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"owner","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/List"}}}}}},"tags":["list"]}},"/api/list/{owner}":{"get":{"operationId":"getByAccountId","parameters":[{"name":"owner","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/List"}}}}}},"tags":["list"]}},"/api/movie":{"post":{"operationId":"add","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddMovieDTO"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Movie"}}}}}},"tags":["movie"]}},"/api/movie/{id}/{listId}":{"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}},{"name":"listId","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Movie"}}}}}},"tags":["movie"]}},"/api/movie/{listId}":{"get":{"operationId":"getByAccountId","parameters":[{"name":"listId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Movie"}}}}}},"tags":["movie"]}}},"info":{"title":"Organize My TV API","description":"A private API to support the Organixe My TV Apps","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UpdateUserDto":{"type":"object","properties":{"accountId":{"type":"number"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"password":{"type":"string"}},"required":["accountId"]},"UpdateUserResponse":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"}},"required":["firstName","lastName","email"]},"SignupDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","minLength":8},"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["email","password","firstName","lastName"]},"Movie":{"type":"object","properties":{"id":{"type":"number"},"movieName":{"type":"string"},"list":{"$ref":"#/components/schemas/List"}},"required":["movieName","list"]},"List":{"type":"object","properties":{"id":{"type":"number"},"listName":{"type":"string"},"movies":{"type":"array","items":{"$ref":"#/components/schemas/Movie"}},"owner":{"$ref":"#/components/schemas/User"}},"required":["listName","movies","owner"]},"User":{"type":"object","properties":{"accountId":{"type":"number"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"lists":{"type":"array","items":{"$ref":"#/components/schemas/List"}},"password":{"type":"string"},"dateCreated":{"type":"string"}},"required":["email","firstName","lastName","lists","password","dateCreated"]},"SignupResponse":{"type":"object","properties":{"token":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["token","user"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","minLength":8}},"required":["email","password"]},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["token","user"]},"ChangePasswordDto":{"type":"object","properties":{"accountId":{"type":"number"},"currentPassword":{"type":"string"},"newPassword":{"type":"string","minLength":8}},"required":["accountId","currentPassword","newPassword"]},"AddListDTO":{"type":"object","properties":{"owner":{"type":"number"},"listName":{"type":"string"}},"required":["owner","listName"]},"AddMovieDTO":{"type":"object","properties":{"list":{"type":"number"},"movieName":{"type":"string"}},"required":["list","movieName"]}}}}